+++ Domain/Entidades/Boletin.cs
@@ -5,7 +5,7 @@ using System.Text;
 
 namespace Domain.Entidades
 {
-    public class Boletin: Entity<int>
+    public class Boletin: Entity<long>
     {
         public List<Nota> ListaNotas { get; private set; }
 
@@ -13,7 +13,7 @@ namespace Domain.Entidades
         {
         }
 
-        public Boletin(int codigoBoletin, List<Nota> notas)
+        public Boletin(long codigoBoletin, List<Nota> notas)
         {
             Id = codigoBoletin;
             ListaNotas = notas;
